% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DendroPCA.R
\name{Dendro_PCA}
\alias{Dendro_PCA}
\title{usa una matriz binaria de individuos en columnas y marcadores en filas y
hace agrupamiento jerarquico y analisis de componentes principales}
\usage{
Dendro_PCA(ruta, archivo, especie)
}
\arguments{
\item{ruta}{es la ruta del archivo de entrada (matriz binaria)}

\item{archivo}{archivo de entrada (matriz binaria)}

\item{especie}{especie de interes de donde vienen los datos}
}
\value{
imagenes de calidad de publicacion de PCA y dendrogramas
calculos de PIC y H y lista de indices cofeneticos
}
\description{
carga el archivo de excel (matriz binaria) y realiza lo siguiente
hace analisis por componentes principales y obtiene grafico de calidad de publicacion.

hace dendrograma con soporte estadistico y obtiene grafico de calidad de publicacion.

a partir de todos (10) los metodos de ade4 para obtener matrices de distancia
y a partir de todos (8) los metodos de agrupamiento jerarquico
obtiene todas las combinaciones y obtiene sus indices cofeneticos
para evaluar los mejores.

contruye dendrogramas para las 4 mejores combinaciones y obtiene grafico de calidad de publicacion.

archivos formato CSV de valores de PIC (con interpretacion) y H

archivo formato CSV de valores cofeneticos para 80 combinaciones de matrices de distancia y
metodos de agrupaiento
}
